+++ src/main/java/service/VoteService.java
@@ -40,10 +40,10 @@ public class VoteService implements IVoteService {
 
     @Override
     public void validate(VoteDTO vote) {
-        int artistId = vote.getArtistId();
+        Long artistId = vote.getArtistId();
         validateArtist(artistId);
 
-        List<Integer> genresIdList = vote.getGenreIds();
+        List<Long> genresIdList = vote.getGenreIds();
         validateGenres(genresIdList);
 
         String about = vote.getAbout();
@@ -53,14 +53,14 @@ public class VoteService implements IVoteService {
         validateEmail(email);
     }
 
-    private void validateArtist(int artistId) {
+    private void validateArtist(Long artistId) {
         if (!artistService.exists(artistId)) {
             throw new NoSuchElementException("Invalid artist id " +
                     "provided - '" + artistId + "' ");
         }
     }
 
-    private void validateGenres(List<Integer> genresIdList) {
+    private void validateGenres(List<Long> genresIdList) {
 
         if (genresIdList.size() < 3 || genresIdList.size() > 5) {
             throw new IllegalArgumentException("Number of genres outside" +
@@ -72,7 +72,7 @@ public class VoteService implements IVoteService {
             throw new IllegalArgumentException("Genre parameter " +
                     "must be non-repeating ");
         }
-        for (int genreId : genresIdList) {
+        for (Long genreId : genresIdList) {
             if (!genreService.exists(genreId)) {
                 throw new NoSuchElementException("Invalid genre id " +
                         "provided - '" + genreId + "' ");
